Move application code to a Python package
https://phabricator.endlessm.com/T34007
