+++ src/kolibri_android/android_whitenoise.py
@@ -5,12 +5,6 @@ import stat
 from urllib.parse import urlparse
 from wsgiref.headers import Headers
 
-from android_utils import document_exists
-from android_utils import document_tree_join
-from android_utils import get_activity
-from android_utils import is_document_uri
-from android_utils import open_file
-from android_utils import stat_file
 from django.contrib.staticfiles import finders
 from django.utils._os import safe_join
 from jnius import autoclass
@@ -25,6 +19,13 @@ from whitenoise.responders import MissingFileError
 from whitenoise.responders import NOT_ALLOWED_RESPONSE
 from whitenoise.responders import Response
 from whitenoise.string_utils import decode_path_info
+
+from .android_utils import document_exists
+from .android_utils import document_tree_join
+from .android_utils import get_activity
+from .android_utils import is_document_uri
+from .android_utils import open_file
+from .android_utils import stat_file
 
 logger = logging.getLogger(__name__)
