+++ src/kolibri_android/initialization.py
@@ -2,15 +2,19 @@ import logging
 import os
 import re
 import sys
+from pathlib import Path
 
-from android_utils import apply_android_workarounds
-from android_utils import get_activity
-from android_utils import get_endless_key_uris
-from android_utils import get_home_folder
-from android_utils import get_signature_key_issuing_organization
-from android_utils import get_timezone_name
-from android_utils import get_version_name
 from jnius import autoclass
+
+from .android_utils import apply_android_workarounds
+from .android_utils import get_activity
+from .android_utils import get_endless_key_uris
+from .android_utils import get_home_folder
+from .android_utils import get_signature_key_issuing_organization
+from .android_utils import get_timezone_name
+from .android_utils import get_version_name
+
+SCRIPT_PATH = Path(__file__).absolute().parent.parent
 
 # initialize logging before loading any third-party modules, as they may cause logging to get configured.
 logging.basicConfig(level=logging.DEBUG)
@@ -19,13 +23,12 @@ jnius_logger.setLevel(logging.INFO)
 
 apply_android_workarounds()
 
-script_dir = os.path.dirname(os.path.abspath(__file__))
-sys.path.append(script_dir)
-sys.path.append(os.path.join(script_dir, "kolibri", "dist"))
-sys.path.append(os.path.join(script_dir, "extra-packages"))
+sys.path.append(SCRIPT_PATH.as_posix())
+sys.path.append(SCRIPT_PATH.joinpath("kolibri", "dist").as_posix())
+sys.path.append(SCRIPT_PATH.joinpath("extra-packages").as_posix())
 
-from android_whitenoise import DynamicWhiteNoise  # noqa: E402
-from android_whitenoise import monkeypatch_whitenoise  # noqa: E402
+from .android_whitenoise import DynamicWhiteNoise  # noqa: E402
+from .android_whitenoise import monkeypatch_whitenoise  # noqa: E402
 
 monkeypatch_whitenoise()
 
@@ -61,14 +64,18 @@ set_content_fallback_dirs_env()
 os.environ["KOLIBRI_APK_VERSION_NAME"] = get_version_name()
 os.environ["DJANGO_SETTINGS_MODULE"] = "kolibri_app_settings"
 
-AUTOPROVISION_FILE = os.path.join(script_dir, "automatic_provision.json")
-if os.path.exists(AUTOPROVISION_FILE):
-    os.environ["KOLIBRI_AUTOMATIC_PROVISION_FILE"] = AUTOPROVISION_FILE
+AUTOPROVISION_PATH = SCRIPT_PATH.joinpath("automatic_provision.json")
+if AUTOPROVISION_PATH.is_file():
+    os.environ["KOLIBRI_AUTOMATIC_PROVISION_FILE"] = AUTOPROVISION_PATH.as_posix()
 
 os.environ["KOLIBRI_CHERRYPY_THREAD_POOL"] = "2"
 
-os.environ["KOLIBRI_APPS_BUNDLE_PATH"] = os.path.join(script_dir, "apps-bundle", "apps")
-os.environ["KOLIBRI_CONTENT_COLLECTIONS_PATH"] = os.path.join(script_dir, "collections")
+os.environ["KOLIBRI_APPS_BUNDLE_PATH"] = SCRIPT_PATH.joinpath(
+    "apps-bundle", "apps"
+).as_posix()
+os.environ["KOLIBRI_CONTENT_COLLECTIONS_PATH"] = SCRIPT_PATH.joinpath(
+    "collections"
+).as_posix()
 
 Secure = autoclass("android.provider.Settings$Secure")
